homework2/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
click==8.3.0
dj-database-url==3.0.1
Django==4.2.11
djangorestframework==3.16.1
gunicorn==23.0.0
gyp==0.1
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.10
pytz==2024.1
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
uvicorn==0.37.0
wheel==0.42.0
whitenoise==6.11.0
